trading-firm/pyproject.toml
[tool.poetry]
name = "trading-firm"
version = "0.1.0"
description = "Raspberry Pi Multi-Agent Trading Desk"
authors = ["Senior Python DevOps Engineer <devops@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
redis = {extras = ["hiredis"], version = "^5.0.4"}
openai = "^1.23.6"
python-telegram-bot = {version = "~=21.0.1", extras = ["job-queue"]}
weaviate-client = "^4.6.1"
pydantic = "^2.7.1"
feedparser = "^6.0.11"
alpaca-trade-api = "^3.2.0"
schedule = "^1.2.1"
pyyaml = "^6.0.1"
# For sentiment analysis in signal_maker
# Note: On a Pi, you might need to build these from source or find ARM64 wheels.
# transformers = "^4.40.1"
# torch = "^2.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
ruff = "^0.4.4"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "C90"]
ignore = []

[tool.black]
line-length = 88
